requirements.txt
Django==5.1.8
celery==5.4.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
google-genai==1.15.0
Pillow==10.4.0
psycopg-binary>=3.2.10
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
